Refactor data_augmentation Classes in another file
